feat: Clear recently viewed
